% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3v5_zoomable_treemap.R
\name{d3v5_zoomable_treemap}
\alias{d3v5_zoomable_treemap}
\alias{d3v5_zoomable_treemap_output}
\alias{render_d3v5_zoomable_treemap}
\title{d3v5 Zoomable Treemap}
\usage{
d3v5_zoomable_treemap(data = flare, width = 800, height = 500,
  elementId = NULL)

d3v5_zoomable_treemap_output(outputId, width = 800, height = 500)

render_d3v5_zoomable_treemap(expr, env = parent.frame(),
  quoted = FALSE)
}
\arguments{
\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a zoomableTreemap}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Implementation of a d3 zoomable treemap

Output and render functions for using zoomableTreemap within Shiny
applications and interactive Rmd documents.
}
