% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_tree.R
\name{df_to_tree}
\alias{df_to_tree}
\title{DF to Tree}
\usage{
df_to_tree(
  df,
  stringPath = "path",
  stringSep = "/",
  value_from = "value",
  root_name = "root"
)
}
\arguments{
\item{df}{A data frame}

\item{stringPath}{The name of the string}

\item{stringSep}{The separator (defaults to "/")}

\item{value_from}{The name of the value (Defaults to "value")}

\item{root_name}{The name of the root (Defaults to "root")}
}
\value{
A list of 2: name and list of n children. The leaf nodes return
  name and value
}
\description{
Convert a data.frame to a list of names and nested children with values
}
\examples{
library(d3tm)
df <- data.frame(path = c("1/A", "1/B/i", "1/B/ii", "2/A", 3), value = 1:5)
df_to_tree(df)
}
