% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flare-data.R
\docType{data}
\name{flare}
\alias{flare}
\title{flare}
\format{An object of class \code{list} of length 2.}
\source{
http://bl.ocks.org/mbostock/4063550
}
\usage{
data(flare)
}
\description{
The flare visualisation toolkit file hierachy
}
\keyword{datasets}
