% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3v5_zoomable_treemap.R
\name{zoomable_treemap_shiny}
\alias{zoomable_treemap_shiny}
\alias{zoomable_treemap_output}
\alias{render_zoomable_treemap}
\title{Shiny bindings for d3v5 zoomableTreemap}
\usage{
zoomable_treemap_output(outputId, width = "100\%", height = 500)

render_zoomable_treemap(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a zoomableTreemap}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using zoomableTreemap within Shiny
applications and interactive Rmd documents.
}
\details{
The name of the clicked node is avaiable based on the object name + _clicked
e.g. input$treemap_clicked
}
