% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoomableTreemap.R
\name{zoomableTreemap}
\alias{zoomableTreemap}
\title{Zoomable Treemap}
\usage{
zoomableTreemap(data, width = 800, height = 500, elementId = NULL)
}
\description{
R Implementation of Mike Bostock's zoomable treemap
}
