% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztm.R
\name{ztm}
\alias{ztm}
\title{Zoomable Treemap}
\usage{
ztm(
  data = jsonlite::read_json(system.file("json/flare-2.json", package = "d3tm")),
  width = "100\%",
  height = 400,
  elementId = NULL
)
}
\description{
A Zoomable Treemap Capable of Handling Unbalanced Hierarchical Data
}
\examples{
if(interactive()) ztm()
}
