% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztm.R
\name{ztmShiny}
\alias{ztmShiny}
\alias{ztmOutput}
\alias{renderZtm}
\title{Shiny Bindings for Zoomable Treemap}
\usage{
ztmOutput(outputId, width = "100\%", height = 500)

renderZtm(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a zoomableTreemap}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using zoomable treemap within Shiny
applications and interactive Rmd documents.
}
\details{
ztm objects send input values to Shiny as the user interacts
with them.

Object input names generally use this pattern:

\code{input$OBJECTID_EVENT_OBJCATEGORY}

So for \code{zoomable_treemap_output("ztm")} clicking on a square would
return the name of the node to \code{input$ztm_clicked_label}.

The full list of current supported events are:

\itemize{
\item \code{input$OBJECTID_clicked_child_id}
\item \code{input$OBJECTID_clicked_child_depth}
\item \code{input$OBJECTID_clicked_child_label}
\item \code{input$OBJECTID_hover_parent_index}
\item \code{input$OBJECTID_hover_parent_depth}
\item \code{input$OBJECTID_hover_parent_label}
\item \code{input$OBJECTID_hover_child_index}
\item \code{input$OBJECTID_hover_child_depth}
\item \code{input$OBJECTID_hover_child_label}
}

If an event has not been triggered it is set to \code{NULL}.

Variables can be reset by sending a custom message in Shiny using:

\code{session$sendCustomMessage("resetInputValue", "OBJECTID_clicked_id")}
}
