% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_nest.R
\name{d3_nest2}
\alias{d3_nest2}
\title{Convert a data.frame to a 'd3.js' Hierarchy}
\usage{
d3_nest2(data = data, value_cols = value_cols, root = root)
}
\arguments{
\item{data}{A data frame input}

\item{value_cols}{The value columns}

\item{root}{The root of the tree name}
}
\value{
A json character vector
}
\description{
Based on the timely portfolio code but with improved performance enhancements
}
\examples{
titanic_df <- data.frame(Titanic)
tit_tb <- titanic_df \%>\%
select(Class,Age,Survived,Sex,Freq) \%>\%
 d3_nest2(value_cols="Freq", root="titanic")
}
