% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoomable_treemap.R
\name{zoomable_treemap}
\alias{zoomable_treemap}
\title{D3 Zoomable Treemap}
\usage{
zoomable_treemap(data = jsonlite::toJSON(flare), width = "100\%",
  height = NULL, elementId = NULL, background = "#bbb",
  header_background = "orange", header_height = 25,
  header_fontsize = "12px", format_string = ",",
  zoom_in_helptext = " - Click on a Square to Zoom",
  zoom_out_helptext = " - Click here to Zoom Out")
}
\arguments{
\item{\code{data}}{A json object}

\item{\code{width}}{Width of the widget, in pixels}

\item{\code{height}}{Height of the widget, in pixels}

\item{\code{elementId}}{Element ID used for R Shiny package}

\item{\code{background}}{Default background colour use if there is no color
variable defined in the data. Takes hex of colour names.}

\item{\code{header_background}}{The background colour either named of in a hex
e.g. "#bbb"}
}
\description{
Implementation of a d3 zoomable treemap
}
\examples{
# download.file("https://raw.githubusercontent.com/d3/d3-hierarchy/master/test/data/flare.json",
#               "inst/examples/flare.json")

json <- jsonlite::fromJSON(system.file("examples/flare.json", package = "d3RZoomableTreemap"))
zoomable_treemap(jsonlite::toJSON(json))
}
