% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoomable_treemap.R
\name{zoomable_treemap}
\alias{zoomable_treemap}
\title{Zoomable Treemap}
\usage{

  zoomable_treemap(data = jsonlite::toJSON(jsonlite::fromJSON(system.file("examples/flare.json",
  package = "d3RZoomableTreemap"))), width = "100\%", height = NULL,
  elementId = NULL, background = "#bbb",
  header_background = "orange", header_height = 25,
  header_fontsize = "12px", format_string = ",",
  zoom_in_helptext = " - Click on a Square to Zoom",
  zoom_out_helptext = " - Click here to Zoom Out",
  tooltip_background = "orange", rmarkdown = FALSE)
}
\arguments{
\item{\code{data}}{A json object}

\item{\code{width}}{Width of the widget, in pixels}

\item{\code{height}}{Height of the widget, in pixels}

\item{\code{elementId}}{Element ID used for R Shiny package}

\item{\code{background}}{Default background colour use if there is no color
variable defined in the data. Takes hex of colour names.}

\item{\code{header_background}}{The background colour either named of in a hex
e.g. "#bbb"}

\item{\code{header_height}}{Height of the title, in pixels}

\item{\code{header_fontsize}}{Font size of the title, in pixels}

\item{\code{format_string}}{D3 number format}

\item{\code{zoom_in_helptext}}{The help text displayed on the header bar}

\item{\code{zoom_out_helptext}}{Zoom out helptext}

\item{\code{tooltip_background}}{Background colour of the tooltip}

\item{\code{rmarkdown}}{Specific whether is an Rmarkdown document, to correct
tooltip location.}
}
\description{
Implementation of a Mike Bostock's d3
 \href{https://bost.ocks.org/mike/treemap/}{zoomable treemap}
 as an htmlwidget in d3v5.
}
\examples{
data.frame(Titanic) \%>\%
 dplyr::select(Class,Age,Survived,Sex,Freq) \%>\%
  d3_nest2(value_col="Freq", root="Titanic",
           id_vars = c("Class", "Age", "Survived", "Sex")) \%>\%
  zoomable_treemap()

}
