% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3v5_zoomable_treemap.R
\name{zoomable_treemap}
\alias{zoomable_treemap}
\title{d3v5 Zoomable Treemap}
\usage{
zoomable_treemap(data = flare, width = 800, height = 500,
  elementId = NULL)
}
\description{
Implementation of a d3 zoomable treemap
}
