% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztm.R
\name{ztm}
\alias{ztm}
\title{<Add Title>}
\usage{
ztm(
  data = jsonlite::read_json(system.file("json/flare-2.json", package = "d3tm")),
  width = "100\%",
  height = 400,
  elementId = NULL
)
}
\description{
<Add Description>
}
