% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztm.R
\name{ztm}
\alias{ztm}
\title{Zoomable Treemap}
\usage{

  ztm(data = jsonlite::toJSON(jsonlite::fromJSON(system.file("examples/flare.json",
  package = "d3RZoomableTreemap"))), width = "100\%", height = NULL,
  elementId = NULL, background = "#bbb",
  header_background = "orange", header_height = 25,
  header_fontsize = "12px", format_string = ",",
  zoom_in_helptext = " - Click on a Square to Zoom",
  zoom_out_helptext = " - Click here to Zoom Out",
  tooltip_background = "orange", rmarkdown = FALSE)
}
\arguments{
\item{data}{A json object}

\item{width}{Width of the widget, in pixels}

\item{height}{Height of the widget, in pixels}

\item{elementId}{Element ID used for R Shiny package}

\item{background}{Default background colour use if there is no color
variable defined in the data. Takes hex of colour names.}

\item{header_background}{The background colour either named of in a hex
e.g. "#bbb"}

\item{header_height}{Height of the title, in pixels}

\item{header_fontsize}{Font size of the title, in pixels}

\item{format_string}{D3 number format}

\item{zoom_in_helptext}{The help text displayed on the header bar}

\item{zoom_out_helptext}{Zoom out helptext}

\item{tooltip_background}{Background colour of the tooltip}

\item{rmarkdown}{Specific whether is an Rmarkdown document, to correct
tooltip location.}
}
\description{
Implementation of a Mike Bostock's d3
 \href{https://bost.ocks.org/mike/treemap/}{zoomable treemap}
 as an htmlwidget in d3v5.
}
\examples{
data.frame(Titanic) \%>\%
 dplyr::select(Class,Age,Survived,Sex,Freq) \%>\%
  d3_nest2(value_col="Freq", root="Titanic",
           id_vars = c("Class", "Age", "Survived", "Sex")) \%>\%
  ztm()

}
